Start untangling glyph_cache from fonts.
