+++ libs/render-wgpu/buffer/font_common/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "glyph_cache"
+name = "font_common"
 version = "0.1.0"
 authors = ["Terrence Cole <terrence.d.cole@gmail.com>"]
 edition = "2018"
@@ -7,7 +7,5 @@ edition = "2018"
 [dependencies]
 failure = "^ 0.1.2"
 image = "^ 0.21"
-lazy_static = ">= 1"
-log = ">= 0.4"
 wgpu = "^ 0.5"
 gpu = { path = "../../../nitrogen/system/gpu" }
