+++ libs/render-wgpu/buffer/glyph_cache/src/glyph_cache.rs
@@ -14,13 +14,10 @@
 // along with OpenFA.  If not, see <http://www.gnu.org/licenses/>.
 use crate::{font_interface::FontInterface, glyph_frame::GlyphFrame};
 
-use codepage_437::{FromCp437, CP437_CONTROL};
-use failure::{ensure, Fallible};
-use fnt::Fnt;
+use failure::Fallible;
 use gpu::GPU;
 use image::{GrayImage, ImageBuffer, Luma};
 use log::trace;
-use rusttype::{Font, Point, Scale};
 use std::collections::HashMap;
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
@@ -38,37 +35,39 @@ impl GlyphCacheIndex {
 
 pub struct GlyphCache {
     index: GlyphCacheIndex,
+    bind_group: wgpu::BindGroup,
     font: Box<dyn FontInterface>,
 }
 
 impl GlyphCache {
-    /*
-    pub fn new_transparent_fnt(
-        fnt: &Fnt,
+    pub fn new(
         index: GlyphCacheIndex,
+        font: Box<dyn FontInterface>,
         bind_group_layout: &wgpu::BindGroupLayout,
-        gpu: &mut GPU,
-    ) -> Fallible<Self> {
-        Ok(Self {
-            index,
-            font: Box::new(FntFont::new_transparent_fnt(fnt, bind_group_layout, gpu)?),
-        })
-    }
+        gpu: &GPU,
+    ) -> Self {
+        let (texture_view, sampler) = font.gpu_resources();
 
-    pub fn new_ttf(
-        bytes: &'static [u8],
-        index: GlyphCacheIndex,
-        bind_group_layout: &wgpu::BindGroupLayout,
-        gpu: &mut GPU,
-    ) -> Fallible<Self> {
-        Ok(Self {
+        let bind_group = gpu.device().create_bind_group(&wgpu::BindGroupDescriptor {
+            label: Some("glyph-cache-TTF-bind-group"),
+            layout: bind_group_layout,
+            bindings: &[
+                wgpu::Binding {
+                    binding: 0,
+                    resource: wgpu::BindingResource::TextureView(&texture_view),
+                },
+                wgpu::Binding {
+                    binding: 1,
+                    resource: wgpu::BindingResource::Sampler(&sampler),
+                },
+            ],
+        });
+
+        Self {
             index,
-            font: Box::new(TtfFont::new_ttf(bytes, bind_group_layout, gpu)?),
-        })
-    }
-     */
-    pub fn new(index: GlyphCacheIndex, font: Box<dyn FontInterface>) -> Self {
-        Self { index, font }
+            bind_group,
+            font,
+        }
     }
 
     pub fn render_height(&self) -> f32 {
@@ -102,7 +101,7 @@ impl GlyphCache {
     }
 
     pub fn bind_group(&self) -> &wgpu::BindGroup {
-        self.font.bind_group()
+        &self.bind_group
     }
 
     pub fn can_render_char(&self, c: char) -> bool {
@@ -115,82 +114,5 @@ impl GlyphCache {
 
     pub fn pair_kerning(&self, a: char, b: char) -> f32 {
         self.font.pair_kerning(a, b)
-    }
-
-    pub fn upload_texture_luma(
-        gpu: &mut GPU,
-        image_buf: ImageBuffer<Luma<u8>, Vec<u8>>,
-    ) -> Fallible<wgpu::TextureView> {
-        let image_dim = image_buf.dimensions();
-        let extent = wgpu::Extent3d {
-            width: image_dim.0,
-            height: image_dim.1,
-            depth: 1,
-        };
-        let image_data = image_buf.into_raw();
-
-        let transfer_buffer = gpu.push_buffer(
-            "glyph-cache-transfer-buffer",
-            &image_data,
-            wgpu::BufferUsage::all(),
-        );
-        let texture = gpu.device().create_texture(&wgpu::TextureDescriptor {
-            label: Some("glyph-cache-texture"),
-            size: extent,
-            array_layer_count: 1,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::R8Unorm,
-            usage: wgpu::TextureUsage::all(),
-        });
-        let mut encoder = gpu
-            .device()
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("glyph-cache-command-encoder"),
-            });
-        encoder.copy_buffer_to_texture(
-            wgpu::BufferCopyView {
-                buffer: &transfer_buffer,
-                offset: 0,
-                bytes_per_row: extent.width,
-                rows_per_image: extent.height,
-            },
-            wgpu::TextureCopyView {
-                texture: &texture,
-                mip_level: 0,
-                array_layer: 0,
-                origin: wgpu::Origin3d::ZERO,
-            },
-            extent,
-        );
-        gpu.queue_mut().submit(&[encoder.finish()]);
-        gpu.device().poll(wgpu::Maintain::Wait);
-
-        let texture_view = texture.create_view(&wgpu::TextureViewDescriptor {
-            format: wgpu::TextureFormat::R8Unorm,
-            dimension: wgpu::TextureViewDimension::D2,
-            aspect: wgpu::TextureAspect::All,
-            base_mip_level: 0,
-            level_count: 1, // mip level
-            base_array_layer: 0,
-            array_layer_count: 1,
-        });
-
-        Ok(texture_view)
-    }
-
-    pub fn make_sampler(device: &wgpu::Device) -> wgpu::Sampler {
-        device.create_sampler(&wgpu::SamplerDescriptor {
-            address_mode_u: wgpu::AddressMode::ClampToEdge,
-            address_mode_v: wgpu::AddressMode::ClampToEdge,
-            address_mode_w: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Nearest,
-            min_filter: wgpu::FilterMode::Nearest,
-            mipmap_filter: wgpu::FilterMode::Nearest,
-            lod_min_clamp: 0f32,
-            lod_max_clamp: 9_999_999f32,
-            compare: wgpu::CompareFunction::Never,
-        })
     }
 }
